requirements.txt
matplotlib
numpy
pandas
pillow
scikit-image
scipy
